deleted superfluous output event beetween challenge and filtering section
